create pool chart tabs provider
