+++ packages/lib/modules/pool/PoolDetail/PoolStats/PoolCharts/usePoolCharts.tsx
@@ -2,7 +2,6 @@ import { ColorMode, theme as defaultTheme, useTheme as useChakraTheme } from '@c
 import { differenceInDays, format } from 'date-fns'
 import {
   GetPoolSnapshotsDocument,
-  GqlPoolType,
   GqlPoolSnapshotDataRange,
   GqlChain,
 } from '@repo/lib/shared/services/api/generated/graphql'
@@ -10,21 +9,13 @@ import { useQuery } from '@apollo/client'
 import { useCallback, useMemo, useState } from 'react'
 import { useParams } from 'next/navigation'
 import { usePool } from '../../../PoolProvider'
-import { PoolVariant, BaseVariant } from '../../../pool.types'
 import { NumberFormatter } from '@repo/lib/shared/utils/numbers'
 import { useCurrency } from '@repo/lib/shared/hooks/useCurrency'
 import { useTheme as useNextTheme } from 'next-themes'
 import { isCowAmmPool } from '../../../pool.helpers'
+import { PoolChartTab, usePoolChartTabs } from './PoolChartTabsProvider'
 
 const MIN_DISPLAY_PERIOD_DAYS = 30
-
-export enum PoolChartTab {
-  VOLUME = 'volume',
-  TVL = 'tvl',
-  FEES = 'fees',
-  SURPLUS = 'surplus',
-  LIQUIDITY_PROFILE = 'liquidity_profile',
-}
 
 export type PoolChartPeriod = {
   value: GqlPoolSnapshotDataRange
@@ -74,17 +65,6 @@ export interface PoolChartTypeOptions {
   areaStyle?: {
     color: ChartColor
   }
-}
-
-type PoolTabsMap = {
-  [key in GqlPoolType]?: PoolChartTypeTab[]
-} & {
-  default: PoolChartTypeTab[] // Make default required
-}
-
-export interface PoolChartTypeTab {
-  value: PoolChartTab
-  label: string
 }
 
 const dataRangeToDaysMap: { [key in GqlPoolSnapshotDataRange]?: number } = {
@@ -196,37 +176,6 @@ export const getDefaultPoolChartOptions = (
   }
 }
 
-const BASE_TABS: PoolChartTypeTab[] = [
-  { value: PoolChartTab.VOLUME, label: 'Volume' },
-  { value: PoolChartTab.TVL, label: 'TVL' },
-]
-
-const POOL_SPECIFIC_TABS: PoolTabsMap = {
-  [GqlPoolType.CowAmm]: [...BASE_TABS, { value: PoolChartTab.SURPLUS, label: 'Surplus' }],
-  [GqlPoolType.Gyroe]: [
-    { value: PoolChartTab.LIQUIDITY_PROFILE, label: 'Liquidity Profile' },
-    ...BASE_TABS,
-    { value: PoolChartTab.FEES, label: 'Fees' },
-  ],
-  default: [...BASE_TABS, { value: PoolChartTab.FEES, label: 'Fees' }],
-}
-
-export function getPoolTabsList({
-  variant,
-  poolType,
-}: {
-  variant: PoolVariant
-  poolType: GqlPoolType
-}): PoolChartTypeTab[] {
-  // LBP pools in v2 only show base tabs
-  if (poolType === GqlPoolType.LiquidityBootstrapping && variant === BaseVariant.v2) {
-    return BASE_TABS
-  }
-
-  // Return pool-specific tabs or default tabs
-  return POOL_SPECIFIC_TABS[poolType] || POOL_SPECIFIC_TABS.default
-}
-
 export function usePoolSnapshots(
   poolId: string,
   chainId: GqlChain,
@@ -246,33 +195,22 @@ export type PoolChartsData = ReturnType<typeof usePoolCharts>
 
 export function usePoolCharts() {
   const { pool, tvl } = usePool()
-  const isCowPool = isCowAmmPool(pool.type)
-
-  const { id: poolId, variant } = useParams()
+  const { id: poolId } = useParams()
   const { toCurrency } = useCurrency()
   const { theme: nextTheme } = useNextTheme()
   const theme = useChakraTheme()
-
-  const tabsList = useMemo(() => {
-    const poolType = pool?.type
-    if (!poolType || typeof variant !== 'string') return []
-
-    return getPoolTabsList({
-      variant: variant as PoolVariant,
-      poolType: poolType,
-    })
-  }, [pool?.type, variant])
-
-  const [activeTab, setActiveTab] = useState(tabsList[0])
   const [chartValue, setChartValue] = useState(0)
   const [chartDate, setChartDate] = useState('')
   const [activePeriod, setActivePeriod] = useState(poolChartPeriods[0])
+  const { activeTab } = usePoolChartTabs()
 
   const { data, loading: isLoadingSnapshots } = usePoolSnapshots(
     poolId as string,
     pool.chain,
     activePeriod.value
   )
+
+  const isCowPool = isCowAmmPool(pool.type)
 
   const chartValueSum = useMemo(() => {
     if (!data?.snapshots) return null
@@ -569,8 +507,6 @@ export function usePoolCharts() {
 
   return {
     isLoading,
-    activeTab,
-    setActiveTab,
     activePeriod,
     setActivePeriod,
     chartValue,
@@ -578,7 +514,6 @@ export function usePoolCharts() {
     options,
     handleAxisMoved,
     chartData,
-    tabsList,
     chartValueSum,
     hasChartData: !!chartData.length,
   }
